% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTry.R
\name{addme}
\alias{addme}
\title{Function which adds 2 two numbers choosen by the user.}
\usage{
addme(a, b)
}
\arguments{
\item{a}{(\code{a}) A number}

\item{b}{(\code{b}) A number}
}
\value{
(\code{numeric}) The addition of the 2 numbers
}
\description{
Function which adds 2 two numbers choosen by the user.
}
\examples{

addme(2, 5)
addme(5, 5)

}
