% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptry.R
\docType{data}
\name{apptry}
\alias{apptry}
\alias{ui}
\title{A simple shiny app that adds 2 two number thnaks to a user interface.}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\value{
(\code{shinyApp}) The web page on which the app is run.
}
\description{
A simple shiny app that adds 2 two number thnaks to a user interface.
}
\keyword{datasets}
