% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator.R
\name{creator}
\alias{creator}
\title{Function which pastes 2 words to create a short sentence.}
\usage{
creator(x, y)
}
\arguments{
\item{x}{(\code{x}) A character/ string}

\item{y}{(\code{y}) A character/ string}
}
\value{
(\code{numeric}) The words group made of the 2 words choosen>
}
\description{
Function which pastes 2 words to create a short sentence.
}
\examples{

creator("Hello", "world")

}
