% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTry-package.R, R/packageTry.R
\docType{package}
\name{packageTry-package}
\alias{packageTry}
\alias{packageTry-package}
\title{packageTry: A first package (trying to be clean)}
\description{
A fisrt simple package containing some basic function, such as add, divide, and some shiny app try

A fisrt simple package containing some basic function, such as add, divide, and some shiny app try
}
\author{
\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)


\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)

}
\keyword{internal}
