% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptry.R
\docType{data}
\name{ui}
\alias{ui}
\title{A simple shiny app which use the function addme to add two numbers thanks to a UI}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\arguments{
\item{ui}{('ui') The modeler of the user interface}

\item{server}{('server') The output part}
}
\value{
A simple app thanks to which the user can add two numbers manualy
}
\description{
A simple shiny app which use the function addme to add two numbers thanks to a UI
}
\keyword{datasets}
