% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptry.R, R/creator.R, R/packageTry-package.R,
%   R/packageTry.R
\docType{package}
\name{packageTry-package}
\alias{packageTry}
\alias{packageTry-package}
\title{packageTry: A first package}
\description{
A first simple package containing some basic functions, such as add, divide, and a shiny app try.

A first simple package containing some basic functions, such as add, divide, and a shiny app try.

A first simple package containing some basic functions, such as add, divide, and a shiny app try.

A first simple package containing some basic functions, such as add, divide, and a shiny app try.
}
\author{
\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)


\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)


\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)


\strong{Maintainer}: Amandine Robert \email{Amandine.Robert@liverpool.ac.uk} (YOUR-ORCID-ID)

}
\keyword{internal}
